CHG: moved mortality and growth declaration from initialisation to the model ADD: growth and mortality pointers to each cell so that they can seed agents with the right parameters everytime they are created

+++ Source/InitialisationPhases/Children/Iterative.h
@@ -48,7 +48,7 @@ protected:
   // members
   unsigned                    years_;
   vector<TimeStep*>           time_steps_;
-  Double                      lambda_;
+  float                      lambda_;
   vector<unsigned>            convergence_years_;
   unsigned                    number_agents_;
   string                      intial_layer_label_;
